+++ bokeh/tests/test_objects.py
@@ -7,7 +7,7 @@ from six.moves import xrange
 
 from bokeh.core.properties import List, String, Instance, Dict, Any, Int
 from bokeh.model import Model
-from bokeh.core.property.containers import PropertyValueList, PropertyValueDict
+from bokeh.core.property.wrappers import PropertyValueList, PropertyValueDict
 from bokeh.util.future import with_metaclass
